+++ CHANGELOG.md
@@ -312,6 +312,7 @@
 
 ### Internals
 * A separate sync version is removed.
+* `SimulatedFailure` mmap handling was not thread-safe.
 
 ----------------------------------------------